% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_primer_elemento.R
\name{mod_primer_elemento_ui}
\alias{mod_primer_elemento_ui}
\alias{mod_primer_elemento_server}
\title{mod_primer_elemento_ui and mod_primer_elemento_server}
\usage{
mod_primer_elemento_ui(id)

mod_primer_elemento_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
